Simplify instantiation of data cache manager
